# Work with a byte slice instead of string