Using `regexp.ReplaceAllString` is very inefficient.  I've refactored the
main clean function to work on a byte slice and called the function `Clean`.
`CleanContent` has been renamed to `CleanString`.

benchmark                 old ns/op     new ns/op     delta
BenchmarkCleanContent     21159375      6637523       -68.63%

benchmark                 old allocs     new allocs     delta
BenchmarkCleanContent     7129           4195           -41.16%

benchmark                 old bytes     new bytes     delta
BenchmarkCleanContent     17182896      598000        -96.52%