@bbalet 2016-01-06
Hi,

Sorry for late answer. Are you sure that byte slices are working with Unicode characters? That was the reason why I chose string instead.