@moorereason 2016-01-06
I'm not well-versed in Unicode.  All of the tests pass.