@bbalet 2016-01-06
Yes I saw that now. I was a bit confused with this point.