@moorereason 2016-01-06
Since you already utilize `golang.org/x/text/unicode/norm`, shouldn't that protect us as long as the stopwords themselves are normalized?