@bbalet 2016-01-06
As far as i undertood, Normalization is a way to uniform the way the glyphs are composed. For example, you can write the glyph 'é' as 'é' or 'e' + put next symbol top + symbol accent. When you normalize a text you set an option according to the form you want to get (canonical or decomposed, NFC or NFD in go).

The problem that might occur is a wrong approach of iteration (e.g. byte by byte would be wrong in my example for the second case). But words are tokenized in my prototype and with the lib simhash, so we iterate on each token and not on each byte.